backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
qdrant-client==1.7.0
cohere==4.34
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pylint==3.0.2
black==23.11.0
mypy==1.7.1